+++ 103 + 309/Project/STM32F10x_StdPeriph_Lib_V3.5.0/drivers/stm32f10x_it.c
@@ -227,6 +227,7 @@ void EXTI9_5_IRQHandler(void)
   }
   if (EXTI_GetITStatus(EXTI_Line9) != RESET)
   {
+    set_irq_wksource((uint8_t)soc_key);
     EXTI_ClearITPendingBit(EXTI_Line9);
   }
 }